requirements.txt
aerich==0.6.3
aiomysql==0.1.1
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
black==22.3.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.1.3
cryptography==37.0.2
dbus-python==1.2.16
dictdiffer==0.9.0
distro-info===0.23ubuntu1
fastapi==0.78.0
h11==0.13.0
idna==2.8
iso8601==1.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
pydantic==1.9.1
PyGObject==3.36.0
PyMySQL==1.0.2
pypika-tortoise==0.1.5
python-apt==2.0.0+ubuntu0.20.4.7
pytz==2022.1
requests==2.22.0
requests-unixsocket==0.2.0
six==1.14.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
tomlkit==0.10.2
tortoise-orm==0.19.1
typing_extensions==4.2.0
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.17.6
